- 完成UI相关代码搬迁
