+++ ui/mainwindow/generated_ui.py
